--- contacts/tests/test_data.html
@@ -1,66 +1,0 @@
-
-
-
-
-
-
-
-
-
-<div id="foiaReports">
-	
-		
-		
-
-
-<div id="foiaReportsTable">
-<div class="breadcrumbcontent">
-<div style="font-weight: normal; font-style: italic; margin: 0 0 10px 0;"><strong>Note:</strong> Please use the links below to move through your previous search results. Your "Back" button will not take you to your previous search results. </div>
-<a href="javascript:breadcrumRequest('0');"><span id="startNewSearch">Start a New Search</span></a>&nbsp;<font color="#999999">&raquo;</font>&nbsp;Processing Time (Advanced)
-</div>
-  
-	<div style="float: right;" id="csvprint">
-	<a href="#" class="foiaHelpText" title="">
-		<img style="float: left; margin: 0px 5px 0pt 5px;" src="images/question-icon.gif" alt="Description of Data from the FOIA.gov Glossary" />
-	</a>
-	<!-- a href="faq.html" class="foiaHelpVideo" title="">
-		<img style="float: left; margin: 0px 10px 0pt 16px;" src="images/video-icon.gif" alt="Description of Data from the FOIA.gov Help Videos" />
-	</a -->
-	      <div class="csvbutton" id="csvbutton"><a class="csvbutton" href="foia/FoiaDownload"><span>CSV</span></a></div>
-	      <div class="printbutton" id="printbutton">&nbsp;&nbsp;&nbsp;&nbsp;<a href="javascript:openPop('foia/FoiaPrint.jsp','_blank');"><img alt="Printer-Friendly Version" src="images/printer-friendly-icon.gif" /></a></div>
-	</div>
-
-<div class="clear"></div>
-<form name="checkedComponents" id="checkedComponents">
-	<div style="float: right; margin: 5px;">
-	    <a class="compareButton" href="javascript:getCheckedComponents(4);">
-	        <span>Compare</span>
-	    </a>  
-	</div>
-	<div class="clear"></div>
-
-	<div style="background:url(images/table-title-bg.jpg) repeat-x bottom left; border: 1px solid #e7e7e7;">
-		<span class="blueTitle">
-		<p>(1) Total Active Results</p></span>
-		<p><a href="javascript:showDiv('agencyInfo0');">Show</a> | <a href="javascript:hideDiv('agencyInfo0');">Hide</a></p>
-	</div>
-	<table border="1" width="100%" cellspacing="0" id="agencyInfo0" style="display:block;">
-	<tr>
-<th><center><a href="javascript:sortRequest('col=1&amp;firstTitle=1&amp;paragraph=0');">Agency<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=2&amp;firstTitle=1&amp;paragraph=0');">Component<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=3&amp;firstTitle=1&amp;paragraph=0');">Year<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=4&amp;firstTitle=1&amp;paragraph=0');">Simple-Median No. of Days<img src="images/ascendingArrows.gif" alt="Sorted Ascending" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=5&amp;firstTitle=1&amp;paragraph=0');">Simple-Average No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=6&amp;firstTitle=1&amp;paragraph=0');">Simple-Lowest No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=7&amp;firstTitle=1&amp;paragraph=0');">Simple-Highest No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=8&amp;firstTitle=1&amp;paragraph=0');">Complex-Median No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=9&amp;firstTitle=1&amp;paragraph=0');">Complex-Average No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=10&amp;firstTitle=1&amp;paragraph=0');">Complex-Lowest No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=11&amp;firstTitle=1&amp;paragraph=0');">Complex-Highest No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=12&amp;firstTitle=1&amp;paragraph=0');">Expedited Processing-Median No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=13&amp;firstTitle=1&amp;paragraph=0');">Expedited Processing-Average No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=14&amp;firstTitle=1&amp;paragraph=0');">Expedited Processing-Lowest No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center><a href="javascript:sortRequest('col=15&amp;firstTitle=1&amp;paragraph=0');">Expedited Processing-Highest No. of Days<img src="images/neutralArrows.gif" alt="Sorting Possible" style="border:0" /></a></center></th><th><center></center></th>
-</tr>
-<tr>
-<td><center><span title="Federal Retirement Thrift Investment Board">FRTIB</span></center></td><td><center><span title="Federal Retirement Thrift Investment Board">FRTIB</span></center></td><td><center>2012</center></td><td><center>20</center></td><td><center>27</center></td><td><center>1</center></td><td><center>57</center></td><td><center>0</center></td><td><center>0</center></td><td><center>0</center></td><td><center>0</center></td><td><center>0</center></td><td><center>0</center></td><td><center>0</center></td><td><center>0</center></td><td><center><input type="checkbox" name="2012560" /></center></td>
-</tr>
-
-	</table>
-	
-	<div style="float: right; margin: 5px;">
-	    <a class="compareButton" href="javascript:getCheckedComponents(4);">
-	        <span>Compare</span>
-	    </a>  
-	</div>
-</form>
-</div>
-	
-	
-</div>
